Update package_config.json and main.dart
