+++ flutter/trokanou1/.dart_tool/package_config.json
@@ -272,7 +272,7 @@
       "languageVersion": "3.2"
     }
   ],
-  "generated": "2024-03-17T13:38:55.524991Z",
+  "generated": "2024-03-17T13:44:14.945289Z",
   "generator": "pub",
   "generatorVersion": "3.3.0"
 }
