+++ flutter/trokanou1/lib/pages/accueil.dart
@@ -1,43 +1,15 @@
 import 'package:flutter/material.dart';
-import 'package:trokannou/pages/usager.dart';
 
-class MyHomePage extends StatefulWidget {
-  const MyHomePage({Key? key}) : super(key: key);
+class Accueil extends StatefulWidget {
+  const Accueil({super.key});
 
   @override
-  State<MyHomePage> createState() => _MyHomePageState();
+  State<Accueil> createState() => _AccueilState();
 }
 
-class _MyHomePageState extends State<MyHomePage> {
-  int _selectedIndex = 0;
-
-  void _navigationBottomBar(int index) {
-    setState(() {
-      _selectedIndex = index;
-    });
-  }
-
-  final List<Widget> _pages = [
-    Usager(),
-    Usager(),
-  ];
-
+class _AccueilState extends State<Accueil> {
   @override
   Widget build(BuildContext context) {
-    return Scaffold(
-      body: _pages[_selectedIndex],
-      bottomNavigationBar: BottomNavigationBar(
-        selectedItemColor: Colors.blue,
-        items: const <BottomNavigationBarItem>[
-          BottomNavigationBarItem(
-              icon: Icon(Icons.home_filled), label: 'Accueil'),
-          BottomNavigationBarItem(icon: Icon(Icons.person), label: 'Usager'),
-          BottomNavigationBarItem(
-              icon: Icon(Icons.add), label: 'Ajouter un produit'),
-        ],
-        currentIndex: _selectedIndex,
-        onTap: _navigationBottomBar,
-      ),
-    );
+    return const Placeholder();
   }
 }
